Destructure Entryform and add EntryForm to FormPage, delete EntryForm
