--- src/components/EntryForm.stories.js
@@ -1,11 +1,0 @@
-import EntryForm from './EntryForm';
-
-export default {
-  title: 'components/EntryForm',
-  component: EntryForm,
-};
-
-const Template = args => <EntryForm {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
