+++ src/pages/FormPage.js
@@ -2,8 +2,33 @@ import { NavLink } from 'react-router-dom';
 import { HiArrowLeft } from 'react-icons/hi';
 import styled from 'styled-components';
 import EntryForm from '../components/EntryForm';
+import SubmitButton from '../components/Button';
+import ScreenRaderOnly from '../components/ScreenRaderOnly';
+import { useState } from 'react';
+
+import Start from '../components/form-pages/Start';
+import Water from '../components/form-pages/Water';
+import Weather from '../components/form-pages/Weather';
+import Catch from '../components/form-pages/Catch';
+import Summary from '../components/form-pages/Summary';
 
 export default function FormPage({ onCreateCard }) {
+  const [formData, setFormData] = useState('');
+
+  const handleOnChange = event => {
+    const { name, value } = event.target;
+    setFormData({
+      ...formData,
+      [name]: value,
+    });
+  };
+
+  function handleSubmit(event) {
+    event.preventDefault();
+    console.log('formData', formData);
+    onCreateCard(formData);
+  }
+
   return (
     <>
       <header>
@@ -13,7 +38,26 @@ export default function FormPage({ onCreateCard }) {
       </header>
       <main>
         <Title>Create a new entry:</Title>
-        <EntryForm onCreateCard={onCreateCard} />
+        <Form
+          aria-labelledby="form-name"
+          onSubmit={handleSubmit}
+          autoComplete="off"
+          labeltext="form"
+        >
+          <Section>General infos</Section>
+          <Start handleOnChange={handleOnChange} />
+          <Section>Water</Section>
+          <Water handleOnChange={handleOnChange} />
+          <Section>Weather</Section>
+          <Weather handleOnChange={handleOnChange} />
+          <Section>Catch</Section>
+          <Catch handleOnChange={handleOnChange} />
+          <Section>Summary</Section>
+          <Summary handleOnChange={handleOnChange} />
+          <SubmitButton text="Submit" isAccent={true} id="form-name">
+            <ScreenRaderOnly>Create your fishing experience</ScreenRaderOnly>
+          </SubmitButton>
+        </Form>
       </main>
     </>
   );
@@ -31,4 +75,17 @@ const LinkStyled = styled(NavLink)`
   align-items: center;
   text-decoration: none;
   color: black;
+`;
+
+const Form = styled.form`
+  padding: 15px 10px 30px;
+  background-color: #fffcf8;
+  color: #a2c36c;
+  border: 0.5px solid #a2c36c;
+  border-radius: 20px;
+  box-shadow: 0px 10px 20px -10px rgba(0, 0, 0, 0.25);
+`;
+
+const Section = styled.span`
+  color: #687a48;
 `;
